Split procs commands into modules
